Add command error output to stderr
This commit appends the error message from the command executed to the ProcessRunner error output. Since the creation and formatting of the error message was duplicated across the code, a private method was created solely for this purpose. This method received the error output as the second parameter, and appends it to the final message.