Changes to ExecutionFailureException to expose error info
This exception class was modified to receive the binary name, the command to be execute and the error output. Then, internally creates and formats the message for the parent constructor. After that, the command and error output variables are stored within the object and each has a getter method. A new test was included to test those getters.